+++ BackEnd/routes/item.js
@@ -1,10 +1,13 @@
 const express = require('express');
 const Item = require('../models/item');
+const User = require('../models/item');
 const router = new express.Router();
+const auth = require('../middleware/auth') 
 const validator = require('../utils/modelsValidate')
 const  definitions  = require('../utils/definitions');
+const Category = require('../models/category');
 
-router.post('/items',async (req, res) => { //create
+router.post('/items',auth,async (req, res) => { //create
     // const item = new Item(({
     //     userId,
     //     title,
@@ -16,17 +19,22 @@ router.post('/items',async (req, res) => { //create
     // } = req.body))
     
     const item = new Item(req.body)
-    console.log(item)
+  console.log(item)
     if (
-        !validator.isValidString(item.userId) ||
         !validator.isValidString(item.title) ||
         !validator.isValidString(item.description) ||
         !validator.isValidString(item.image) ||
         !validator.isValidString(item.price) ||
-        !validator.isValidString(item.category) ||
+        !validator.isValidValue(item.category) ||
         !validator.isValidString(item.quantity) 
     ) { //check empty fields 
         res.status(500).send(definitions.MISSING_FIELDS);
+        return;
+    }
+    console.log(item.category)
+    const category = await Category.findById(item.category)
+    if (!category) {
+        res.status(500).send(definitions.CATEGORY_NOT_FOUND);
         return;
     }
     if (!validator.isValidPrice(item.price)) { //check total price is positive 
@@ -38,11 +46,11 @@ router.post('/items',async (req, res) => { //create
     }
     
     await item.save()
-    console.log(item)
+    // console.log(item)
     return res.status(201).send({item})
     
 });
-router.get('/getOneItem/:id', async (req, res) => { //getOne
+router.get('/getOneItem/:id',auth, async (req, res) => { //getOne
     // const item_id = (req.params.id)
 
     // Item.findById(item_id).then(item => {
@@ -51,20 +59,15 @@ router.get('/getOneItem/:id', async (req, res) => { //getOne
     //     }
     //     return res.status(200).send(item);
     // })
-
-    const allItems = await Item.find({});
-    let item;
-    let flag = false
-    allItems.find(element => {
-            if ((element._id == req.params.id)) {
-                item = element;
-                flag = true
-            }
-        });
-    if (flag) {
-        return res.status(200).send(item);
+    if (!(req.params.id.match(/^[0-9a-fA-F]{24}$/))) { //check if this a valid ObjectId
+        return res.status(404).send(`The item ${req.params.id} not found!`);
     }
-    return res.status(500).send(`The item ${req.params.id} not found!`);
+    try {
+        const findItem = await Item.findById(req.params.id);
+        return res.status(200).send(findItem);
+    } catch (e) {
+        return res.status(500).send();
+    }
 })
 router.get('/getAllItems', async (req, res) => {//getAll
    
@@ -79,6 +82,9 @@ router.delete('/items/:id', async (req, res) => {//delete
     const item_id = req.params.id;
     
     try {
+        if (!(item_id.match(/^[0-9a-fA-F]{24}$/))) { //check if this a valid ObjectId
+            return res.status(404).send({ message: 'Item not found' });
+        }
         const item = await Item.findByIdAndDelete(item_id)
         if (!item) {
             return res.status(404).send('No items are found!');
@@ -90,21 +96,69 @@ router.delete('/items/:id', async (req, res) => {//delete
         }
 });
 router.patch('/item/:id', async (req, res) => { //update
-    const updates = Object.keys(req.body)
-    const allowedUpdates = ['userId','title', 'description','image','total','stock']
-    const isValidOperation = updates.every((update) => {
-        allowedUpdates.includes(update)
-    })
-    if (isValidOperation) {
-        return res.status(400).send({ error: 'Invalid updates!' })
-    }
+    const itemId = req.params.id;
+    const { title, description, image, price, quantity } = req.body;
     try {
-        updates.forEach((update) => (req.item[update] = req.body[update]))
-        await req.item.save()
-        res.send(req.item)
-    } catch (e) {
-        return res.statusMessage(500).send(e)
-    }
+        if (!(itemId.match(/^[0-9a-fA-F]{24}$/))) { //check if this a valid ObjectId
+            return res.status(404).send({ message: 'Item not found' });
+        }
+        const itemToUpdate = await Item.findById(itemId);
+
+        if (title !== undefined && title !== itemToUpdate.title ) {
+            if (validator.isValidString(title)) {
+                itemToUpdate.title = title
+            } else {
+                res.status(500).send(definitions.MISSING_FIELDS);
+                return;
+            }
+        }
+        if (description !== undefined && description !== itemToUpdate.description ) {
+            if (validator.isValidString(description)) {
+                itemToUpdate.description = description
+            } else {
+                res.status(500).send(definitions.MISSING_FIELDS);
+                return;
+            }
+        }
+        if (image !== undefined && image !== itemToUpdate.image ) {
+            if (validator.isValidString(image)) {
+                itemToUpdate.image = image
+            } else {
+                res.status(500).send(definitions.MISSING_FIELDS);
+                return;
+            }
+        }
+        if (price !== undefined && price !== itemToUpdate.price) {
+            if (!(validator.isValidString(price))) {
+                res.status(500).send(definitions.MISSING_FIELDS);
+                return;
+                
+            } else if (!(validator.isValidPrice(price))) {
+                res.status(500).send(definitions.PRICE_NOT_VALID);
+                return;
+            }
+            itemToUpdate.price = price
+
+        }
+        if (quantity !== undefined && quantity !== itemToUpdate.quantity) {
+                if (!(validator.isValidString(quantity))) {
+                    res.status(500).send(definitions.MISSING_FIELDS);
+                    return;
+                    
+                } else if (!(validator.isAvailable(quantity))) {
+                    res.status(500).send(definitions.QUANTITY_NOT_AVAILABLE);
+                    return;
+                }
+                itemToUpdate.quantity = quantity
+            }
+        // Find the item by ID and update the specified fields
+        // const item = await Item.findByIdAndUpdate(itemId, updatedFields, { new: true });
+        await itemToUpdate.save();
+        return res.send({ message: 'Item updated successfully', itemToUpdate });
+      } catch (error) {
+        console.error(error);
+        res.status(500).send({ message: 'Server error' });
+      }
 })
