+++ BackEnd/routes/user.js
@@ -4,11 +4,12 @@ const router = new express.Router();
 const auth = require('../middleware/auth') 
 const bcrypt = require('bcryptjs')
 const validator = require('../utils/modelsValidate')
-const  definitions  = require('../utils/definitions');
+const definitions = require('../utils/definitions');
+const jwt = require('jsonwebtoken')
 //CRUD - create read update delete
 
 
-router.post('/users',async (req, res) =>  {
+router.post('/users',auth,async (req, res) =>  {
     const user = new User(req.body);
     if (
         !validator.isValidString(user.userName) ||
@@ -22,64 +23,42 @@ router.post('/users',async (req, res) =>  {
         res.status(500).send(definitions.MISSING_FIELDS);
         return;
     }
+    
+    const { userName, email } = req.body;
+    
+    const isExistingUserName = await User.findOne({ userName });
+    if (isExistingUserName) {
+        res.status(500).send(definitions.USER_EXIST);
+        return;
+    }
+
     if (!validator.isValidEmail(user.email)) {//check email is valid
         res.status(500).send(definitions.EMAIL_NOT_VALID);
         return;
     }
+
+    const isExistingEmail = await User.findOne({ email });
+    if (isExistingEmail) {
+        res.status(500).send(definitions.EMAIL_EXIST);
+        return;
+    }
+
     if (!validator.isValidPhoneNumber(user.phoneNumber)) { //check phone number is contains 10 digits
         res.status(500).send(definitions.PHONE_NUMBER_NOT_VALID);
         return;
     }
+
     if (!validator.isValidPassword(user.password)) { // check password is more than 6 characters
         res.status(500).send(definitions.PASSWORD_NOT_VALID);
     }
-
+    const token = jwt.sign({ _id: user._id.toString() }, "thisismyshoppingcart")
+    console.log(token)
+    user.tokens = user.tokens.concat({ token })
     await user.save()
     console.log(user)
     return res.status(201).send({user})
 })
-
-router.post('/userss',async (req, res) =>  { //create
-    
-    // const me = new User({
-    //     userName: 'Shai1712',
-    //     firstName: 'shai',
-    //     lastName: 'Shais',
-    //     password: 4325248,
-    //     address: 'NYC22',
-    //     phoneNumber: '052376545',
-    //     email: 'Shai22112@gmail.com'
-    // });
-    const user  = new User(req.body)
-    if (
-        !validator.isValidString(user.userName) ||
-        !validator.isValidString(user.firstName) ||
-        !validator.isValidString(user.lastName) ||
-        !validator.isValidString(user.password) ||
-        !validator.isValidString(user.address) ||
-        !validator.isValidString(user.phoneNumber) ||
-        !validator.isValidString(user.email)
-    ) { //check empty fields 
-        res.status(500).send(definitions.MISSING_FIELDS);
-        return;
-    }
-    if (!validator.isValidEmail(user.email)) {//check email is valid
-        res.status(500).send(definitions.EMAIL_NOT_VALID);
-        return;
-    }
-    if (!validator.isValidPhoneNumber(user.phoneNumber)) { //check phone number is contains 10 digits
-        res.status(500).send(definitions.PHONE_NUMBER_NOT_VALID);
-        return;
-    }
-    if (!validator.isValidPassword(user.password)) { // check password is more than 6 characters
-        res.status(500).send(definitions.PASSWORD_NOT_VALID);
-    }
-
-    await user.save()
-    console.log(user)
-    return res.status(201).send({user})
-});
-router.post('/users/login',auth, async (req, res) => {
+router.post('/users/login', async (req, res) => {
     const { userName, password } = req.body;
     let result;
     let message;
@@ -88,14 +67,14 @@ router.post('/users/login',auth, async (req, res) => {
         const user = await User.findOne({ userName })
         if (!user) {
             message = "User is not exists!"
-            result = ({user})
+            result = []
           
         }
         else {
             const isMatch = await bcrypt.compare(password, user.password)
             if (!isMatch) {
                 message = "Invalid password!"
-                result =({ user})
+                result = []
                
             } else {
                 message = "Login successfully!"
@@ -106,26 +85,46 @@ router.post('/users/login',auth, async (req, res) => {
         }
     } else {
         message = "Missing fields!"
+        result = []
     }
     return res.send({message,result})
 })
-router.get('/getOneUser/:id', async (req, res) => { //getOne
-    const allUsers = await User.find({});
-    let user;
-    let flag = false
-    allUsers.find(element => {
-            if ((element._id == req.params.id)) {
-                user = element;
-                flag = true
-            }
-        });
-    if (flag) {
-        return res.status(200).send(user);
+
+router.post('/users/logout', auth, async (req, res) => {
+    try {
+        req.user.tokens = req.user.tokens.filter((token) => {
+            return token.token !== req.token
+        })
+        await req.user.save()
+        return res.status(200).send(`The user has logout successfully!!`);
+    } catch (e) {
+        return res.status(500).send(`Something went wrong!!`)
     }
-    return res.status(500).send(`The user ${req.params.id} not found!`);
+    
+})
+// router.post('/users/logoutAll', auth, async (req, res) => {
+//     try {
+//         req.user.tokens = []
+//         await req.user.save()
+//         return res.send();
+//     } catch (e) {
+//         return res.status(500).send()
+//     }
+// })
+router.get('/getProfileUser/:id',auth, async (req, res) => { //getOne
+
+    if (!(req.params.id.match(/^[0-9a-fA-F]{24}$/))) { //check if this a valid ObjectId
+        return res.status(404).send(`The user ${req.params.id} not found!`);
+    }
+    try {
+        const findUser = await User.findById(req.params.id);
+        return res.status(200).send(findUser);
+    } catch (e) {
+        return res.status(500).send();
+    }
 
     });
-    router.get('/getAllUsers', async (req, res) => { //getAll
+    router.get('/getAllUsers',auth, async (req, res) => { //getAll
         const users = await User.find({});
     
         if (users == null || users.length === 0) {
@@ -133,9 +132,12 @@ router.get('/getOneUser/:id', async (req, res) => { //getOne
         }
         return res.status(201).send(users) 
     });
-router.delete('/users/:id', async (req, res) => { //delete
+router.delete('/users/:id',auth, async (req, res) => { //delete
     const user_id = req.params.id;
     try {
+        if (!(user_id.match(/^[0-9a-fA-F]{24}$/))) { //check if this a valid ObjectId
+            return res.status(404).send({ message: 'ObjectId not found' });
+        }
         const user = await User.findByIdAndDelete(user_id)
         if (!user) {
             return res.status(404).send('No users are found!');
@@ -148,36 +150,86 @@ router.delete('/users/:id', async (req, res) => { //delete
     });
 
     router.patch('/users/:id',async (req, res) => { //update
-        // const updates = [(req.body)];
-        // console.log(updates)
-        // const allowedUpdates = [('firstName', 'lastName', 'address', 'phoneNumber')];
-        // const isValidOperation = updates.every(update => {
-        //     console.log(update)
-        //     allowedUpdates.includes([update]);
-        // });
-        // console.log(isValidOperation)
-        // if (!isValidOperation) {
-        //     return res.status(400).send({ error: 'Invalid updates!' });
-        // }
-        // try {
-        //     updates.forEach((update) => (req.user[update] = req.body[update]));
-        //     await req.user.save();
-        //     res.send(req.user);
-        // } catch (e) {
-        //     return res.status(500).send(e);
-        // }
+       
         const userId = req.params.id;
-        const updatedFields = req.body;
-
+        const {userName,firstname,lastName,password,address,phoneNumber,email} = req.body;
         try {
-            // Find the user by ID and update the specified fields
-            const user = await User.findByIdAndUpdate(userId, updatedFields, { new: true });
-        
-            if (!user) {
-              return res.status(404).send({ message: 'User not found' });
+            if (!(userId.match(/^[0-9a-fA-F]{24}$/))) { //check if this a valid ObjectId
+                return res.status(404).send({ message: 'User not found' });
             }
-        
-            res.json({ message: 'User updated successfully', user });
+
+            // Find the user by ID and update the specified fields
+            const userToUpdate = await User.findById(userId);
+            if (userId !== userToUpdate.id) {
+                if (validator.isValidValue(userId)) {
+                    userToUpdate.id = userId
+                }
+            }
+            if (userName !== userToUpdate.userName) {
+          
+                if (validator.isValidString(userName)) {
+                    userToUpdate.userName = userName
+                }else {
+                    res.status(500).send(definitions.MISSING_FIELDS);
+                    return;
+                }
+            }
+            if (firstname !== undefined && firstname !== userToUpdate.first ) {
+                if (validator.isValidString(firstname)) {
+                    userToUpdate.firstName = firstname
+                } else {
+                    res.status(500).send(definitions.MISSING_FIELDS);
+                    return;
+                }
+            }
+            if (lastName !== userToUpdate.lastName) {
+                if (validator.isValidString(lastName)) {
+                    userToUpdate.lastName = lastName
+                }else {
+                    res.status(500).send(definitions.MISSING_FIELDS);
+                    return;
+                }
+            }
+          
+            const isMatch = await bcrypt.compare(password, userToUpdate.password)
+            if (!isMatch) {
+                res.status(500).send(definitions.PASSWORD_CANNOT_CHANGE);
+                return;
+        }
+            if (address !== userToUpdate.address) {
+                if (validator.isValidString(address)) {
+                    userToUpdate.address = address
+                }else {
+                    res.status(500).send(definitions.MISSING_FIELDS);
+                    return;
+                }
+            }
+            if (phoneNumber !== userToUpdate.phoneNumber) {
+                if (!(validator.isValidString(phoneNumber))) {
+                    res.status(500).send(definitions.MISSING_FIELDS);
+                    return;
+                }else if(!(validator.isValidPhoneNumber(phoneNumber))){
+                    res.status(500).send(definitions.PHONE_NUMBER_NOT_VALID);
+                    return;
+                }
+                userToUpdate.phoneNumber = phoneNumber
+            }
+            if (email !== userToUpdate.email) {
+                if (!(validator.isValidString(email))) {
+                    res.status(500).send(definitions.MISSING_FIELDS);
+                    return;
+                    
+                } else if (!(validator.isValidEmail(email))) {
+                    res.status(500).send(definitions.EMAIL_NOT_VALID);
+                    return;
+                }
+                userToUpdate.email = email
+            }
+           
+           // const updatedUser = await User.findByIdAndUpdate(userId, updatedFields, { new: true });
+            await userToUpdate.save();
+            return res.send({ message: 'User updated successfully', userToUpdate });
+            // return res.send({ message: 'User updated successfully' });
           } catch (error) {
             console.error(error);
             res.status(500).send({ message: 'Server error' });
